refactor: useCalculateCommandListHeight hook으로 분리
